+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Drop support for PHP 5.5 (#29)
 
 ### Fixed
+- Update the content element wizard for TYPO3 8.7 (#42)
 - Fix the redirect_url on TravisCI on TYPO3 8.7 (#40)
 - Make the PHPUnit test runner configurable (#37)